+++ CHANGELOG.md
@@ -1,3 +1,4 @@
+- fix: renaming bug
 # v0.3.0
 - add: creation time display
 - add: sort by creation time