+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 - fix: renaming bug
+- improve: change the deletion behavior from direct deletion to move to recycle bin
 # v0.3.0
 - add: creation time display
 - add: sort by creation time